fm_radio/test_launcher.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Power API</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body/>

  <script type="application/javascript">
  // It is a requirement that power_basics.html is in the directory path that is noted below. 
  // If the file is not present then the test will fail silently

  var url = window.location.href.replace(/:\/\/[^\/]*\//, "://example.org/");
  ok (true, url);
  url = url.substring(0, url.lastIndexOf('/')) + "/fm_basics.html"; 
  var domain = "http://example.org/";
  ok(true, url);
  SimpleTest.waitForExplicitFinish();
  SpecialPowers.addPermission("mozFM", true, "http://example.org/");
  SpecialPowers.addPermission("fmradio", true, "http://example.org/");
  SpecialPowers.addPermission("settings", true, "http://example.org/");

  var iframe = document.createElement("iframe");
  iframe.mozApp = "http://example.org/manifest.webapp";
  iframe.src = url;
  iframe.setAttribute("id", "ifrm");
  iframe.onload = loaded();
  document.body.appendChild(iframe);

  function loaded() {

    window.addEventListener("message", function onMessage(event) {
       if (event.data == "finished") {
         window.removeEventListener("message", onMessage, false);
         SimpleTest.finish();
       } else {
         dump(JSON.stringify(event.data));
         ok(event.data.result, event.data.msg);
       }
           
     }, false);

  }
  </script>
</head>

</html>
